Implementation of the new security layer that authenticated based off of a CEVO API
This is quite complex:

1) Listener looks for the cookie to have been set, sets CEVOToken with that
2) Authentication provider uses the sessionid set on that cookie - calls out to the ApiManager
3) ApiManager uses the sessionid to request the current user's information, passes back to Authentication Provider
4) AuthenticationProvider finds or creates the user based on the information. Then creates an AuthenticatedToken, sets the user on it, and we're done!

The Authentication Listener also watches for the cookie to disappear, meaning that we're logged out.

In addition, there is a small part of the system that represents a fake CEVO page. In development environment, we're sent to those pages both in the browser and for API. This allows us to fake everything.