Adding 2 interfaces that describe what's needed when we link one entity (e.g. GamePage) to many locales.
This is accomplished by creating what looks like a ManyToMany relationship with a join table (e.g. game_page_locale). However, since locale isn't a table, it's actually a ManyToOne relationship from GamePage to a real entity called GamePageLocale.

This has all been done in the past, but these 2 interfaces were created so that the magic of the LocalesRelationshipListener class could be re-used in the future. The purpose of this listener is to guarantee that we can call GamePage->getLocales() and it will return an array of the string locales that are joined via the GamePage.gamePageLocales relationship.

We will use this to join the Deal entity to many locales